requirements.txt
discord.py
kwslogger
PyYAML
aiosqlite